Refactored Intelligent Decision Engine for automated tool selection and parameter optimization
- IntelligentDecisionEngine class to analyze target profiles and select optimal tools based on effectiveness ratings.
- Methods for technology detection, attack surface calculation, and risk level determination.
- Advanced parameter optimization for various security tools.
- Comprehensive attack chain creation process tailored to target types and objectives.
- Detailed tool effectiveness ratings and attack patterns for various scenarios.
