+++ core/parameter_optimizer.py
@@ -1,0 +1,245 @@
+from typing import Any, Dict, List, Optional
+from datetime import datetime
+from core.technology_detector import TechnologyDetector
+from core.rate_limit_detector import RateLimitDetector
+from core.failure_recovery_system import FailureRecoverySystem
+from core.performance_monitor import PerformanceMonitor
+from shared.target_profile import TargetProfile
+
+class ParameterOptimizer:
+    """Advanced parameter optimization system with intelligent context-aware selection"""
+
+    def __init__(self):
+        self.tech_detector = TechnologyDetector()
+        self.rate_limiter = RateLimitDetector()
+        self.failure_recovery = FailureRecoverySystem()
+        self.performance_monitor = PerformanceMonitor()
+
+        # Tool-specific optimization profiles
+        self.optimization_profiles = {
+            "nmap": {
+                "stealth": {
+                    "scan_type": "-sS",
+                    "timing": "-T2",
+                    "additional_args": "--max-retries 1 --host-timeout 300s"
+                },
+                "normal": {
+                    "scan_type": "-sS -sV",
+                    "timing": "-T4",
+                    "additional_args": "--max-retries 2"
+                },
+                "aggressive": {
+                    "scan_type": "-sS -sV -sC -O",
+                    "timing": "-T5",
+                    "additional_args": "--max-retries 3 --min-rate 1000"
+                }
+            },
+            "gobuster": {
+                "stealth": {
+                    "threads": 5,
+                    "delay": "1s",
+                    "timeout": "30s"
+                },
+                "normal": {
+                    "threads": 20,
+                    "delay": "0s",
+                    "timeout": "10s"
+                },
+                "aggressive": {
+                    "threads": 50,
+                    "delay": "0s",
+                    "timeout": "5s"
+                }
+            },
+            "sqlmap": {
+                "stealth": {
+                    "level": 1,
+                    "risk": 1,
+                    "threads": 1,
+                    "delay": 1
+                },
+                "normal": {
+                    "level": 2,
+                    "risk": 2,
+                    "threads": 5,
+                    "delay": 0
+                },
+                "aggressive": {
+                    "level": 3,
+                    "risk": 3,
+                    "threads": 10,
+                    "delay": 0
+                }
+            }
+        }
+
+    def optimize_parameters_advanced(self, tool: str, target_profile: TargetProfile, context: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
+        """Advanced parameter optimization with full intelligence"""
+        if context is None:
+            context = {}
+
+        # Get base parameters
+        base_params = self._get_base_parameters(tool, target_profile)
+
+        # Detect technologies for context-aware optimization
+        detected_tech = self.tech_detector.detect_technologies(
+            target_profile.target,
+            headers=context.get("headers", {}),
+            content=context.get("content", ""),
+            ports=target_profile.open_ports
+        )
+
+        # Apply technology-specific optimizations
+        tech_optimized_params = self._apply_technology_optimizations(tool, base_params, detected_tech)
+
+        # Monitor system resources and optimize accordingly
+        resource_usage = self.performance_monitor.monitor_system_resources()
+        resource_optimized_params = self.performance_monitor.optimize_based_on_resources(tech_optimized_params, resource_usage)
+
+        # Apply profile-based optimizations
+        profile = context.get("optimization_profile", "normal")
+        profile_optimized_params = self._apply_profile_optimizations(tool, resource_optimized_params, profile)
+
+        # Add metadata
+        profile_optimized_params["_optimization_metadata"] = {
+            "detected_technologies": detected_tech,
+            "resource_usage": resource_usage,
+            "optimization_profile": profile,
+            "optimizations_applied": resource_optimized_params.get("_optimizations_applied", []),
+            "timestamp": datetime.now().isoformat()
+        }
+
+        return profile_optimized_params
+
+    def _get_base_parameters(self, tool: str, profile: TargetProfile) -> Dict[str, Any]:
+        """Get base parameters for a tool"""
+        base_params = {"target": profile.target}
+
+        # Tool-specific base parameters
+        if tool == "nmap":
+            base_params.update({
+                "scan_type": "-sS",
+                "ports": "1-1000",
+                "timing": "-T4"
+            })
+        elif tool == "gobuster":
+            base_params.update({
+                "mode": "dir",
+                "threads": 20,
+                "wordlist": "/usr/share/wordlists/dirbuster/directory-list-2.3-medium.txt"
+            })
+        elif tool == "sqlmap":
+            base_params.update({
+                "batch": True,
+                "level": 1,
+                "risk": 1
+            })
+        elif tool == "nuclei":
+            base_params.update({
+                "severity": "critical,high,medium",
+                "threads": 25
+            })
+
+        return base_params
+
+    def _apply_technology_optimizations(self, tool: str, params: Dict[str, Any], detected_tech: Dict[str, List[str]]) -> Dict[str, Any]:
+        """Apply technology-specific optimizations"""
+        optimized_params = params.copy()
+
+        # Web server optimizations
+        if "apache" in detected_tech.get("web_servers", []):
+            if tool == "gobuster":
+                optimized_params["extensions"] = "php,html,txt,xml,conf"
+            elif tool == "nuclei":
+                optimized_params["tags"] = optimized_params.get("tags", "") + ",apache"
+
+        elif "nginx" in detected_tech.get("web_servers", []):
+            if tool == "gobuster":
+                optimized_params["extensions"] = "php,html,txt,json,conf"
+            elif tool == "nuclei":
+                optimized_params["tags"] = optimized_params.get("tags", "") + ",nginx"
+
+        # CMS optimizations
+        if "wordpress" in detected_tech.get("cms", []):
+            if tool == "gobuster":
+                optimized_params["extensions"] = "php,html,txt,xml"
+                optimized_params["additional_paths"] = "/wp-content/,/wp-admin/,/wp-includes/"
+            elif tool == "nuclei":
+                optimized_params["tags"] = optimized_params.get("tags", "") + ",wordpress"
+            elif tool == "wpscan":
+                optimized_params["enumerate"] = "ap,at,cb,dbe"
+
+        # Language-specific optimizations
+        if "php" in detected_tech.get("languages", []):
+            if tool == "gobuster":
+                optimized_params["extensions"] = "php,php3,php4,php5,phtml,html"
+            elif tool == "sqlmap":
+                optimized_params["dbms"] = "mysql"
+
+        elif "dotnet" in detected_tech.get("languages", []):
+            if tool == "gobuster":
+                optimized_params["extensions"] = "aspx,asp,html,txt"
+            elif tool == "sqlmap":
+                optimized_params["dbms"] = "mssql"
+
+        # Security feature adaptations
+        if detected_tech.get("security", []):
+            # WAF detected - use stealth mode
+            if any(waf in detected_tech["security"] for waf in ["cloudflare", "incapsula", "sucuri"]):
+                optimized_params["_stealth_mode"] = True
+                if tool == "gobuster":
+                    optimized_params["threads"] = min(optimized_params.get("threads", 20), 5)
+                    optimized_params["delay"] = "2s"
+                elif tool == "sqlmap":
+                    optimized_params["delay"] = 2
+                    optimized_params["randomize"] = True
+
+        return optimized_params
+
+    def _apply_profile_optimizations(self, tool: str, params: Dict[str, Any], profile: str) -> Dict[str, Any]:
+        """Apply optimization profile settings"""
+        if tool not in self.optimization_profiles:
+            return params
+
+        profile_settings = self.optimization_profiles[tool].get(profile, {})
+        optimized_params = params.copy()
+
+        # Apply profile-specific settings
+        for key, value in profile_settings.items():
+            optimized_params[key] = value
+
+        # Handle stealth mode flag
+        if params.get("_stealth_mode", False) and profile != "stealth":
+            # Force stealth settings even if different profile requested
+            stealth_settings = self.optimization_profiles[tool].get("stealth", {})
+            for key, value in stealth_settings.items():
+                optimized_params[key] = value
+
+        return optimized_params
+
+    def handle_tool_failure(self, tool: str, error_output: str, exit_code: int, current_params: Dict[str, Any]) -> Dict[str, Any]:
+        """Handle tool failure and suggest recovery"""
+        failure_analysis = self.failure_recovery.analyze_failure(error_output, exit_code)
+
+        recovery_plan = {
+            "original_tool": tool,
+            "failure_analysis": failure_analysis,
+            "recovery_actions": [],
+            "alternative_tools": failure_analysis["alternative_tools"],
+            "adjusted_parameters": current_params.copy()
+        }
+
+        # Apply automatic parameter adjustments based on failure type
+        if failure_analysis["failure_type"] == "timeout":
+            if "timeout" in recovery_plan["adjusted_parameters"]:
+                recovery_plan["adjusted_parameters"]["timeout"] *= 2
+            if "threads" in recovery_plan["adjusted_parameters"]:
+                recovery_plan["adjusted_parameters"]["threads"] = max(1, recovery_plan["adjusted_parameters"]["threads"] // 2)
+            recovery_plan["recovery_actions"].append("Increased timeout and reduced threads")
+
+        elif failure_analysis["failure_type"] == "rate_limited":
+            timing_profile = self.rate_limiter.adjust_timing(recovery_plan["adjusted_parameters"], "stealth")
+            recovery_plan["adjusted_parameters"].update(timing_profile)
+            recovery_plan["recovery_actions"].append("Applied stealth timing profile")
+
+        return recovery_plan
