+++ core/performance_monitor.py
@@ -1,0 +1,96 @@
+import time
+import psutil
+from typing import Dict, Any
+import logging
+logger = logging.getLogger(__name__)
+
+class PerformanceMonitor:
+    """Advanced performance monitoring with automatic resource allocation"""
+
+    def __init__(self):
+        self.performance_metrics = {}
+        self.resource_thresholds = {
+            "cpu_high": 80.0,
+            "memory_high": 85.0,
+            "disk_high": 90.0,
+            "network_high": 80.0
+        }
+
+        self.optimization_rules = {
+            "high_cpu": {
+                "reduce_threads": 0.5,
+                "increase_delay": 2.0,
+                "enable_nice": True
+            },
+            "high_memory": {
+                "reduce_batch_size": 0.6,
+                "enable_streaming": True,
+                "clear_cache": True
+            },
+            "high_disk": {
+                "reduce_output_verbosity": True,
+                "enable_compression": True,
+                "cleanup_temp_files": True
+            },
+            "high_network": {
+                "reduce_concurrent_connections": 0.7,
+                "increase_timeout": 1.5,
+                "enable_connection_pooling": True
+            }
+        }
+
+    def monitor_system_resources(self) -> Dict[str, float]:
+        """Monitor current system resource usage"""
+        try:
+            cpu_percent = psutil.cpu_percent(interval=1)
+            memory = psutil.virtual_memory()
+            disk = psutil.disk_usage('/')
+            network = psutil.net_io_counters()
+
+            return {
+                "cpu_percent": cpu_percent,
+                "memory_percent": memory.percent,
+                "disk_percent": disk.percent,
+                "network_bytes_sent": network.bytes_sent,
+                "network_bytes_recv": network.bytes_recv,
+                "timestamp": time.time()
+            }
+        except Exception as e:
+            logger.error(f"Error monitoring system resources: {str(e)}")
+            return {}
+
+    def optimize_based_on_resources(self, current_params: Dict[str, Any], resource_usage: Dict[str, float]) -> Dict[str, Any]:
+        """Optimize parameters based on current resource usage"""
+        optimized_params = current_params.copy()
+        optimizations_applied = []
+
+        # CPU optimization
+        if resource_usage.get("cpu_percent", 0) > self.resource_thresholds["cpu_high"]:
+            if "threads" in optimized_params:
+                original_threads = optimized_params["threads"]
+                optimized_params["threads"] = max(1, int(original_threads * self.optimization_rules["high_cpu"]["reduce_threads"]))
+                optimizations_applied.append(f"Reduced threads from {original_threads} to {optimized_params['threads']}")
+
+            if "delay" in optimized_params:
+                original_delay = optimized_params.get("delay", 0)
+                optimized_params["delay"] = original_delay * self.optimization_rules["high_cpu"]["increase_delay"]
+                optimizations_applied.append(f"Increased delay to {optimized_params['delay']}")
+
+        # Memory optimization
+        if resource_usage.get("memory_percent", 0) > self.resource_thresholds["memory_high"]:
+            if "batch_size" in optimized_params:
+                original_batch = optimized_params["batch_size"]
+                optimized_params["batch_size"] = max(1, int(original_batch * self.optimization_rules["high_memory"]["reduce_batch_size"]))
+                optimizations_applied.append(f"Reduced batch size from {original_batch} to {optimized_params['batch_size']}")
+
+        # Network optimization
+        if "network_bytes_sent" in resource_usage:
+            # Simple heuristic for high network usage
+            if resource_usage["network_bytes_sent"] > 1000000:  # 1MB/s
+                if "concurrent_connections" in optimized_params:
+                    original_conn = optimized_params["concurrent_connections"]
+                    optimized_params["concurrent_connections"] = max(1, int(original_conn * self.optimization_rules["high_network"]["reduce_concurrent_connections"]))
+                    optimizations_applied.append(f"Reduced concurrent connections to {optimized_params['concurrent_connections']}")
+
+        optimized_params["_optimizations_applied"] = optimizations_applied
+        return optimized_params
